support deactivation in language server mode
Track the resources created by the extension so that it can be deactivated cleanly. Uses the pre-3.0 backcompat technique from https://docs.sourcegraph.com/extensions/authoring/activation#backcompat-for-sourcegraph-versions-prior-to-3-0 to avoid throwing an error on pre-3.0 versions of Sourcegraph.

This does not clean up the resources created in search mode (using basic code intel), but it's a start.