show Go ifaces/impls location provider in a panel (#47)
Previously, the call to `registerLocationProvider('id', ...` was a noop. The location provider was not actually used anywhere. It needs to be added to a panel to be shown.

- Depends on https://github.com/sourcegraph/sourcegraph/pull/2733
- Depends on #46 

![goifaceimpls](https://user-images.githubusercontent.com/1976/54329595-86957780-45cf-11e9-91fe-17f0b2392290.png)

Unlike for TypeScript (see note in https://github.com/sourcegraph/sourcegraph-typescript/pull/132), Go ifaces/impls were already broken, so this should be merged immediately even though https://github.com/sourcegraph/sourcegraph/pull/2733 is not yet merged and released.